*net uip_addr_xxx  deploed to net/ip/uipaddr.h. This need to avoid deps for all uip resources.
* standalone uipaddr.h allows to use uip_addr structures without all uip API. need by uip addr printers later
